docker error

+++ node/api/user/users.js
@@ -3,7 +3,7 @@ const User = require('../../models/user');
 const Comment = require('../../models/comment');
 const bcrypt = require('bcrypt-nodejs');
 
-
+//get('/users')
 const readAll = async (req, res, next) => {
   const users = await User.findAll().catch((err) => {
     console.error(err);
@@ -12,8 +12,9 @@ const readAll = async (req, res, next) => {
   return res.json(users);
 }
 
+//post('/users')
 const register = async (req, res, next) => {
-  const { loginId, pw, name, age, married } = req.body;
+  const { loginId, pw, name, age, married, isAdmin } = req.body;
 
   const result = await User.count({
     where: {
@@ -42,7 +43,8 @@ const register = async (req, res, next) => {
               pw: bcrypt.hashSync(pw),
               name,
               age,
-              married
+              married,
+              isAdmin
             }).catch((e) => {
               console.error(e);
               return next(e);
@@ -55,10 +57,11 @@ const register = async (req, res, next) => {
   }
 }
 
+//get('/users/:id')
 const read = async (req, res, next) => {
   const id = req.params.id;
   const user = await User.findOne({
-    where: { id: id },
+    where: { loginId: id, deletedAt: null },
   }).catch((err) => {
     console.error(err);
     return next(err);
@@ -68,13 +71,19 @@ const read = async (req, res, next) => {
   return response(res, 200, user);
 }
 
+//patch('/users/:id')
 const update = async (req, res, next) => {
+  const loginId = req.user.dataValues.loginId; //현재 로그인된 아이디
+  const id = req.params.id;
+  
+  if (loginId != id && !req.user.dataValues.isAdmin) return response(res, 403, 'No Authentication');
+
   const result = await User.update({
     name: req.body.name,
     age: req.body.age,
     married: req.body.married,
   }, {
-    where: { id: req.params.id },
+    where: { loginId: id },
   }).catch((err) => {
     console.error(err);
     return next(err);
@@ -84,7 +93,13 @@ const update = async (req, res, next) => {
   return response(res, 200, result);
 }
 
+//delete('/users/:id')
 const remove = async (req, res, next) => {
+  const loginId = req.user.dataValues.loginId; //현재 로그인된 아이디
+  const id = req.params.id;
+  
+  if (loginId != id && !req.user.dataValues.isAdmin) return response(res, 403, 'No Authentication');
+  
   const result = User.destroy({
     where: { loginId: id },
   }).catch((e) => {
@@ -96,18 +111,39 @@ const remove = async (req, res, next) => {
   return response(res, 200);
 }
 
+//get('/users/logout')
 const logout = async (req, res, next) => {
   req.logout();
   req.session.destroy();
   res.redirect('/login');
 }
 
+//get('/users/:id/comments')
 const commentRead = async (req, res, next) => {
+  const id = req.params.id;
+
+  //존재하는 유저인지 확인
+  const user = await User.findOne({
+    where: {
+      [Op.and]: [
+        { id: id },
+        { deletedAt: null }
+      ]
+    }
+  });
+  if (!user) return response(res, 400, 'cannot find user');
+
+
   const comments = await Comment.findAll({
     include: {
       model: User,
-      where: { id: req.params.id },
-    },
+      where: {
+        [Op.and]: [
+          { id: id },
+          { parentId: null }
+        ]
+      }
+    }
   }).catch((err) => {
     console.error(err);
     return next(err);
